requirements.txt
apstone==0.0.8
audioread==3.0.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.0
cv2box==0.5.9
cycler==0.12.1
decorator==5.2.1
einops==0.8.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
msgpack==1.1.0
networkx==3.2.1
numba==0.60.0
numexpr==2.10.2
numpy==1.23.5
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime-gpu==1.16.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==6.30.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2024.8.30
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typeguard==2.13.3
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
